+++ src/app/(application)/profile/page.tsx
@@ -1,0 +1,19 @@
+import { Body, Title } from '@/components/ui/Page';
+
+import { getCurrentUser } from '@/utils/auth/session';
+import { redirect } from 'next/navigation';
+
+export default async function Loading(): Promise<JSX.Element> {
+  const user = await getCurrentUser();
+
+  if (!user) {
+    redirect('/auth/signin');
+  }
+
+  return (
+    <>
+      <Title heading="Profile" subtitle="Customize your profile." />
+      <Body>Profile</Body>
+    </>
+  );
+}