+++ src/components/ui/Input.tsx
@@ -1,20 +1,28 @@
-import { cn } from '@/utils/cn';
+export type InputProps = React.InputHTMLAttributes<HTMLInputElement> & {
+  helpertext?: string;
+  error?: string;
+};
 
 import * as React from 'react';
 
-export interface InputProps extends React.InputHTMLAttributes<HTMLInputElement> {}
+import { cn } from '@/utils/cn';
 
-const Input = React.forwardRef<HTMLInputElement, InputProps>(({ className, type, ...props }, ref) => {
+const Input = React.forwardRef<HTMLInputElement, InputProps>(({ className, type, error, helpertext, ...props }, ref) => {
   return (
-    <input
-      type={type}
-      className={cn(
-        'flex h-9 w-full rounded-md border border-input bg-transparent px-3 py-1 text-sm shadow-sm transition-colors file:border-0 file:bg-transparent file:text-sm file:font-medium placeholder:text-muted-foreground focus-visible:outline-none focus-visible:ring-1 focus-visible:ring-ring disabled:cursor-not-allowed disabled:opacity-50',
-        className
-      )}
-      ref={ref}
-      {...props}
-    />
+    <>
+      <input
+        type={type}
+        className={cn(
+          'flex h-10 w-full rounded-md border border-input bg-transparent px-3 py-2 text-sm ring-offset-background file:border-0 file:bg-transparent file:text-sm file:font-medium placeholder:text-muted-foreground focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2 disabled:cursor-not-allowed disabled:opacity-50',
+          className
+        )}
+        ref={ref}
+        {...props}
+      />
+      {helpertext || error ? (
+        <p className={cn('px-1 text-xs', error ? 'text-destructive' : 'text-muted-foreground')}>{helpertext || error}</p>
+      ) : undefined}
+    </>
   );
 });
 Input.displayName = 'Input';